+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2021-12-20
+
+### Added
+
+- Implemented ability to do force stop by timeout
+
+### Updated
+
+- Updated `configuration`, tests
+- Updated `server`, tests
+- Updated `main`, tests
+- Updated consts, package documentation
+
 ## [1.3.5] - 2021-12-16
 
 ### Updated